Deletion: Employee moved to scheduler.models
