+++ src/main/java/com/iwm/backend/schedulegenerator/util/CalculationsUtility.java
@@ -1,6 +1,6 @@
 package com.iwm.backend.schedulegenerator.util;
 
-import com.iwm.backend.api.models.Employee;
+import com.iwm.backend.schedulegenerator.models.Employee;
 import com.iwm.backend.api.models.Shift;
 
 import java.util.HashMap;
